requirements.txt
asgiref==3.6.0
DateTime==5.1
defusedxml==0.7.1
diff-match-patch==20200713
Django==4.1.7
django-cors-headers==3.14.0
django-cors-middleware==1.5.0
django-import-export==3.1.0
django-phonenumber-field==7.0.2
django-phonenumbers==1.0.1
django-widget-tweaks==1.4.12
et-xmlfile==1.1.0
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.1.2
phonenumbers==8.13.7
pytz==2022.7.1
PyYAML==6.0
sqlparse==0.4.3
tablib==3.3.0
Unidecode==1.3.6
xlrd==2.0.1
xlwt==1.3.0
zope.interface==6.0
